--- src/models/ExampleUser.js
@@ -1,15 +1,0 @@
-import mongoose from "mongoose";
-const { Schema } = mongoose;
-
-// First we create our schema
-const ExampleUserSchema = new Schema({
-    username: { type: String, unique: true },
-    // The schemas can have very flexible type definition options, see https://mongoosejs.com/docs/schematypes.html
-    age: { type: Number, default: 420 },
-});
-
-// Then we enable using it by converting it into a model
-const ExampleUser = mongoose.model("ExampleUser", ExampleUserSchema);
-
-// Which we can then export
-export default ExampleUser;
